+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/PlayQueue.kt
@@ -1,8 +1,10 @@
 package ac.mdiq.podcini.storage.model
 
 import io.realm.kotlin.ext.realmListOf
+import io.realm.kotlin.ext.realmSetOf
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.RealmSet
 import io.realm.kotlin.types.annotations.Ignore
 import io.realm.kotlin.types.annotations.PrimaryKey
 import java.util.*
@@ -21,6 +23,11 @@ class PlayQueue : RealmObject {
 
     @Ignore
     val episodes: MutableList<Episode> = mutableListOf()
+
+    var idsBin: RealmSet<Long> = realmSetOf()
+
+//    @Ignore
+//    val episodesBin: MutableList<Episode> = mutableListOf()
 
     fun isInQueue(episode: Episode): Boolean {
         return episodeIds.contains(episode.id)